Soccer/requirements.txt
apscheduler==3.10.4
fastapi==0.118.0
joblib==1.5.2
numpy==2.3.2
pandas==2.3.2
lightgbm==4.6.0
catboost==1.2.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
scikit-learn==1.7.1
streamlit==1.49.1
uvicorn==0.37.0
